refactor(task_runners): carve the package into seven focused submodules
``orchestration/task_runners/__init__.py`` was a 1355-LOC god module
holding four task-run modes plus their shared machinery.  Split along
the seams the audit identified — no logic changes, only relocation:

- container.py — podman launch + lifecycle primitives (_run_container,
                 _podman_start, _assert_running, _agent_runner,
                 _project_runtime_flags, _print_login_instructions).
- shield.py    — per-task egress-firewall policy (_apply_shield_policy
                 and its read/write/restore/drop/auth-protect helpers).
- config.py    — agent-config assembly + unrestricted-mode helpers.
- cli.py       — task_run_cli (interactive mode).
- toad.py      — task_run_toad + the Caddy token/URL/resume helpers.
- headless.py  — task_run_headless / task_followup_headless plus the
                 HeadlessRunRequest / DetachedSummary value objects.
- restart.py   — task_restart.

Dependency DAG (no cycles): {container, shield, config} are the
foundation; cli / toad / headless build on them; restart additionally
reuses toad's token/URL helpers.

``__init__.py`` is now a pure re-export hub — the seven public names
(task_run_{cli,toad,headless}, task_followup_headless, task_restart,
HeadlessRunRequest, DetachedSummary) still import from
``terok.lib.orchestration.task_runners``, so api.py, domain/task.py and
domain/project.py are untouched.

Tests: four files updated.  Private-symbol imports re-pointed at the
owning submodule; patch targets for moved internals redirected to the
module that actually performs the lookup — e.g. ``_agent_runner`` /
``has_gpu`` → ``task_runners.container``, ``_apply_shield_policy`` /
``_shield_down_impl`` / ``get_shield_bypass_firewall_no_protection`` →
``task_runners.shield``, the ``_resume_toad_container`` test patches →
``task_runners.toad`` (where that runner resolves its imports).

All 2301 unit tests pass; lint, tach, lint-imports, docstrings (97.4%),
REUSE all green.
